requirements.txt
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.18.0
